Slowly breaking structs.h apart
