+++ compiler.cpp
@@ -19,6 +19,7 @@
 //#include "win32_helpers.h"
 #include "structs.h"
 #include "memory.h"
+#include "audio.h"
 #include "plugin.h"
 #include "compiler.h"
