+++ main.cpp
@@ -9,14 +9,14 @@
 #include "base.h"
 #include "structs.h"
 #include "memory.h"
-#include "plugin.h"
 #include "audio.h"
+#include "plugin.h"
 #include "win32_helpers.h"
 #include "wav_reader.h"
+#include "draw.h"
 #include "font.h"
 #include "dsp.h"
 #include "app.h"
-#include "draw.h"
 
 #include "win32_platform.h"
 #include "opengl.h"
