+++ memory.h
@@ -21,6 +21,14 @@
 //} Memory_Log;
 //
 
+
+struct Arena{
+    char *base;
+    char *current; 
+    u64 capacity;
+}; 
+
+
 #define __FILENAME__ (strrchr(__FILE__, '\\') ? strrchr(__FILE__, '\\') + 1 : __FILE__)
 
 //#define LOG_ALLOCATIONS 1
@@ -84,10 +92,10 @@ internal void *arena_allocate(Arena *allocator, u64 size)
     {
         allocator->current += aligned;
         
-        printf("%f\n", float(allocator->current - allocator->base) / float(allocator->capacity));
         return begin;
     }
     else{
+        printf("%f\n", float(allocator->current - allocator->base) / float(allocator->capacity));
         ensure(false);
         return nullptr;
     }
