+++ structs.h
@@ -3,12 +3,6 @@
 #ifndef STRUCTS_H
 #define STRUCTS_H
 
-struct Arena{
-    char *base;
-    char *current; 
-    u64 capacity;
-}; 
-
 //~ Assets
 
 enum Asset_File_State : u32 {
@@ -109,162 +103,6 @@ struct Compiler_Gui_Log{
 };
 
 
-//~ Audio File
-
-
-enum Wav_Reading_Error{
-    Wav_Success,
-    Wav_Could_Not_Open_File,
-    Wav_Not_A_RIFF,
-    Wav_File_Reading_Error,
-    Wav_Invalid_Format
-};
-
-typedef struct {
-    Wav_Reading_Error error;
-    u32 num_channels;
-    u32 samples_by_channel;
-    real32** deinterleaved_buffer;
-    u32 read_cursor;
-} Audio_File;
-
-
-//~ Plugin
-
-typedef void(*audio_callback_t)(void*, void*, float**, unsigned int, unsigned int, float);
-typedef void(*default_parameters_t)(void*);
-typedef void(*initialize_state_t)(void*, void*, unsigned int, float, void*);
-
-
-enum  Plugin_Parameter_Type{
-    Int,
-    Float,
-    Enum
-};
-
-
-typedef struct{
-    i32 min;
-    i32 max;
-} Parameter_Int;
-
-typedef struct {
-    real32 min;
-    real32 max;
-    bool log;
-} Parameter_Float;
-
-typedef struct{
-    i64 value;
-    String name;
-} Parameter_Enum_Entry;
-
-
-typedef struct {
-    Parameter_Enum_Entry *entries;
-    u32 num_entries;
-} Parameter_Enum;
-
-typedef struct {
-    Custom_Error error;
-    String name;
-    u32 offset;
-    
-    Plugin_Parameter_Type type;
-    union {
-        Parameter_Int int_param;
-        Parameter_Float float_param;
-        Parameter_Enum enum_param;
-    };
-    
-} Plugin_Descriptor_Parameter;
-
-typedef struct {
-    String name;
-    //le nom c'est le filename ?
-    Custom_Error error;
-    struct { 
-        i64 size;
-        i64 alignment;
-    } parameters_struct;
-    
-    struct { 
-        i64 size;
-        i64 alignment;
-    } state_struct;
-    
-    Plugin_Descriptor_Parameter *parameters;
-    u32 num_parameters;
-} Plugin_Descriptor;
-
-typedef struct {
-    union{
-        int int_value;
-        float float_value;
-        int enum_value;
-    };
-} Plugin_Parameter_Value;
-
-
-//TODO pour l'instant il y a un bug : si l'ui arrive à emettre 4096 blocs pendant que le thread audio fait 1 copie
-typedef struct  {
-    Plugin_Parameter_Value *buffer;
-    volatile Plugin_Parameter_Value **head;
-    u32 writer_idx;
-    u32 buffer_size;
-    u32 num_fields_by_plugin;
-} Plugin_Parameters_Ring_Buffer;
-
-
-typedef struct Plugin {
-    Compiler_Failure_Stage failure_stage;
-    Clang_Error_Log clang_error_log;
-    Decl_Search_Log decls_search_log;
-    
-    Plugin_Descriptor descriptor;
-    
-    void* llvm_jit_engine;
-    audio_callback_t audio_callback_f;
-    default_parameters_t default_parameters_f;
-    initialize_state_t initialize_state_f;
-    
-    char* parameters_holder;
-    char* state_holder;
-    
-    Plugin_Parameter_Value* parameter_values_audio_side;
-    Plugin_Parameter_Value* parameter_values_ui_side;
-    Plugin_Parameters_Ring_Buffer ring;
-} Plugin;
-
-typedef struct{
-    Arena *allocator;
-} Plugin_Initialization_Context;
-
-
-//~ Audio
-
-typedef struct 
-{
-    real32 sample_rate;
-    u32 num_channels;
-    u32 num_samples;
-    u32 bit_depth;
-} Audio_Parameters;
-
-struct Plugin_Reloading_Manager;
-
-typedef struct 
-{
-    i8 audio_file_play;
-    i8 audio_file_loop;
-    i8 plugin_play;
-    
-    Asset_File_State *plugin_state;
-    Asset_File_State *audio_file_state;
-    
-    Plugin_Reloading_Manager *m;
-    Audio_File *audio_file;
-} Audio_Thread_Context;
 
 //~ UI
 
@@ -312,82 +150,6 @@ typedef struct{
 } UI_State;
 
 struct Window_Context;
-
-
-
-
-//~ Fonts
-
-//TODO types 
-typedef struct {
-    real32 X0, Y0, X1, Y1;     // Glyph corners
-    real32 U0, V0, U1, V1;     // Texture coordinates
-    i32 advance_x;
-    
-    u32 codepoint; 
-} Glyph;
-
-typedef struct{
-    float font_size; 
-    
-    i32 *codepoint_to_idx;  
-    real32 *codepoint_to_advancex; 
-    u32 highest_codepoint;
-    
-    Glyph *glyphs;
-    u32 glyph_count;
-    
-    Vec2 white_rect_pos;
-    //TODO fallback
-    
-    real32 ascent;
-    real32 descent;
-    
-    u32* atlas_pixels;
-    Vec2 atlas_texture_dim;
-    u32 atlas_texture_id;
-} Font;
-
-//~ Graphics
-
-typedef struct {
-    Vec2 pos;
-    Vec2 uv;
-    Color col;
-} Vertex;
-
-typedef struct{
-    Vec2 pos;
-    Vec2 quad_pos;
-} IR_Vertex;
-
-typedef struct {
-    Font font;
-    Vertex *draw_vertices;
-    u32 draw_vertices_count;
-    u32 *draw_indices;
-    u32 draw_indices_count;
-} Graphics_Context_Atlas;
-
-typedef struct {
-    Rect bounds;
-    real32 *IR_buffer;
-    u32 IR_sample_count;
-    real32 zoom_state;
-} Graphics_Context_IR;
-
-typedef struct {
-    Rect bounds;
-    real32 *fft_buffer;
-    u32 fft_sample_count;
-} Graphics_Context_FFT;
-
-typedef struct {
-    Vec2 window_dim;
-    Graphics_Context_Atlas atlas;
-    Graphics_Context_IR ir;
-    Graphics_Context_FFT fft;
-} Graphics_Context;
 
 //~ DSP/IR/FFT
