+++ wasapi_audio.cpp
@@ -13,8 +13,8 @@
 #include "base.h"
 #include "structs.h"
 #include "memory.h"
-#include "plugin.h"
 #include "audio.h"
+#include "plugin.h"
 
 const IID IID_IAudioClient  = __uuidof(IAudioClient);
 const IID IID_IAudioRenderClient = __uuidof(IAudioRenderClient);
